test(utils): split out transform helper
