+++ packages/utils/src/lib/utils.ts
@@ -34,43 +34,6 @@ export const detailsTableHeaders: string[] = [
   'Line(s)',
 ];
 
-// === Transform
-
-export function pluralize(text: string): string {
-  if (text.endsWith('y')) {
-    return text.slice(0, -1) + 'ies';
-  }
-  if (text.endsWith('s')) {
-    return `${text}es`;
-  }
-  return `${text}s`;
-}
-
-export function toArray<T>(val: T | T[]): T[] {
-  return Array.isArray(val) ? val : [val];
-}
-
-export function objectToKeys<T extends object>(obj: T) {
-  return Object.keys(obj) as (keyof T)[];
-}
-
-export function objectToEntries<T extends object>(obj: T) {
-  return Object.entries(obj) as [keyof T, T[keyof T]][];
-}
-
-export function countOccurrences<T extends PropertyKey>(
-  values: T[],
-): Partial<Record<T, number>> {
-  return values.reduce<Partial<Record<T, number>>>(
-    (acc, value) => ({ ...acc, [value]: (acc[value] ?? 0) + 1 }),
-    {},
-  );
-}
-
-export function distinct<T extends string | number | boolean>(array: T[]): T[] {
-  return Array.from(new Set(array));
-}
-
 // @TODO move to report.ts
 export function formatReportScore(score: number): string {
   return Math.round(score * 100).toString();
